updated css placement, and created slider-save.js
put internal css inside of the <head> element at the top of the page.

Created a slider-save.js file to save the vlaue of the slider into the local storage in case that the page refreshes, also goes back to the default value of 5 when they clear the forms content with the model in the settings.